added macro sha256str, regex_struct_name
sha256str takes an expression and returns the sha256 hex string back.
Expression must be AsRef<[u8]>.

regex_struct_name takes an expression and returns the name for that
regex subexpression, which is the leftmost half of a sha256 hex string
prefix with two double scores. The naming is to avoid potential clashes
as well as being able to reuse expressions.